implement data service
